requirements.txt
future==0.18.3
iso8601==2.1.0
pymodbus==3.5.4
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyserial==3.5
PyYAML==6.0.1
